# Conflicts:
#	CMakeLists.txt
#	Common/ac/gamesetupstructbase.cpp
#	Common/ac/gamestructdefines.h
#	Common/core/def_version.h
#	Common/game/room_file.cpp
#	Common/game/roomstruct.cpp
#	Editor/AGS.Editor/Resources/agsdefns.sh
#	Editor/AGS.Editor/app.manifest
#	Editor/AGS.Types/Enums/ScriptAPIVersion.cs
#	Editor/AGS.Types/Properties/AssemblyInfo.cs
#	Editor/AGS.Types/Room.cs
#	Engine/ac/characterextras.cpp
#	Engine/ac/draw.cpp
#	Engine/ac/room.cpp
#	Engine/ac/roomstatus.cpp
#	Engine/ac/roomstatus.h
#	Engine/game/game_init.cpp
#	Engine/game/savegame_v321.cpp
#	Engine/gfx/ali3dogl.h
#	Engine/gfx/ali3dsw.h
#	Engine/gfx/ddb.h
#	Engine/main/engine_setup.cpp
#	Engine/platform/windows/gfx/ali3dd3d.h
#	version.json

+++ Engine/ac/characterextras.cpp
@@ -89,8 +89,8 @@ void CharacterExtras::WriteToSavegame(Stream *out) const
     out->WriteInt8(slow_move_counter);
     out->WriteInt16(animwait);
     // since version 2
-    out->WriteInt8(anim_volume);
-    out->WriteInt8(cur_anim_volume);
+    out->WriteInt8(static_cast<uint8_t>(anim_volume));
+    out->WriteInt8(static_cast<uint8_t>(cur_anim_volume));
     out->WriteInt8(0); // reserved to fill int32
     out->WriteInt8(0);
     // since version 10
